requirements.txt
APScheduler==3.6.3
argcomplete==2.0.0
attrs==22.1.0
awscli==1.24.8
black==22.3.0
blinker==1.4
botocore==1.26.8
Brlapi==0.8.4
cached-property==1.5.2
cachetools==4.2.2
certifi==2022.6.15
chardet==4.0.0
click==8.0.3
colorama==0.4.5
command-not-found==0.3
cryptography==3.4.8
crytic-compile==0.2.4
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
filelock==3.7.0
Flask==2.2.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.6.0
keyring==23.8.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.22.0
MarkupSafe==2.1.1
more-itertools==8.10.0
mypy-extensions==0.4.3
netifaces==0.11.0
numpy==1.23.3
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.2.0
pipx==1.0.0
platformdirs==2.5.2
prettytable==3.5.0
profanity-check==1.0.3
psycopg2==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
Pygments==2.12.0
PyGObject==3.42.2
PyJWT==2.4.0
pyparsing==3.0.7
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.1
pysha3==1.0.2
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian==0.1.46
python-dotenv==0.20.0
python-telegram==0.15.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.11
requests==2.27.1
roman==3.3
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
SecretStorage==3.3.3
six==1.16.0
slither-analyzer==0.9.1
solc-select==1.0.1
systemd-python==235
texttable==1.6.4
tg @ file:///home/lucascullen/GitHub/tg
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
tzdata==2022.6
tzlocal==4.2
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
userpath==1.8.0
virtualenv==20.16.3+ds
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.2.3
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
zipp==1.0.0
